[FIX] remove port check

+++ scripts/common.sh
@@ -390,9 +390,11 @@ function Exit_Clear() {
 
 # check python urllib3 for aliyun (CentOS 7.x)
 Check_Python_Urllib(){
-    if ( which pip > /dev/null 2>&1 );then
-        if ( pip show urllib3 > /dev/null 2>&1 );then
-            pip uninstall urllib3 -y  > /dev/null 2>&1 
+    if [ ! -f $INIT_FILE ];then
+        if ( which pip > /dev/null 2>&1 );then
+            if ( pip show urllib3 > /dev/null 2>&1 );then
+                pip uninstall urllib3 -y  > /dev/null 2>&1 
+            fi
         fi
     fi
 }
